Cleanup: Remove unnecessary appsock handlers, convert some to plugins
appsock handlers removed:
* explain
* genid48
* masterhost
* repopnewlrl
* testcompr
* version

appsock handlers converted to plugins:
* alias
* logdelete
* remsql
