+++ db/CMakeLists.txt
@@ -32,7 +32,7 @@ set(src
   lrucache.c
   marshal.c
   memdebug.c
-  mtrap.c
+  #mtrap.c
   osql_srs.c
   osqlblkseq.c
   osqlblockproc.c
