+++ db/comdb2.h
@@ -2971,7 +2971,6 @@ BB_COMPILE_TIME_ASSERT(make_node_incoherent_req,
     ((rc) == IX_FND || (rc) == IX_NOTFND || (rc) == IX_FNDMORE ||              \
      (rc) == IX_PASTEOF)
 
-void mtrap_init(void);
 void mtrap(char *msg);
 
 int get_file_lwm(int lwmp[2]);
@@ -3401,17 +3400,11 @@ extern int __slow_write_ns;
 
 #include "dbglog.h"
 
-#include "comdb2_appsock.h"
-int appsock_repopnewlrl(comdb2_appsock_arg_t *arg);
-int handle_partition(comdb2_appsock_arg_t *arg);
-
 void handle_testcompr(SBUF2 *sb, const char *table);
 void handle_setcompr(SBUF2 *);
 void handle_rowlocks_enable(SBUF2 *);
 void handle_rowlocks_enable_master_only(SBUF2 *);
 void handle_rowlocks_disable(SBUF2 *);
-void handle_genid48_enable(SBUF2 *);
-void handle_genid48_disable(SBUF2 *);
 
 extern int gbl_bbipc_slotidx;
 
@@ -3614,5 +3607,7 @@ int rename_table_options(void *tran, struct dbtable *db, const char *newname);
 
 int comdb2_get_verify_remote_schemas(void);
 void comdb2_set_verify_remote_schemas(void);
+
+int repopulate_lrl(const char *p_lrl_fname_out);
 
 #endif /* !INCLUDED_COMDB2_H */
