+++ db/fdb_comm.c
@@ -1,5 +1,5 @@
 /*
-   Copyright 2015 Bloomberg Finance L.P.
+   Copyright 2015, 2018 Bloomberg Finance L.P.
 
    Licensed under the Apache License, Version 2.0 (the "License");
    you may not use this file except in compliance with the License.
@@ -26,260 +26,13 @@
 #include "ssl_bend.h" /* for gbl_client_ssl_mode & gbl_ssl_allow_remsql */
 
 extern int gbl_fdb_track;
-extern int gbl_fdb_track_times;
 extern int gbl_time_fdb;
 extern int gbl_notimeouts;
-
 extern int gbl_expressions_indexes;
+
 void free_cached_idx(uint8_t **cached_idx);
 
-/* matches fdb_svc_callback_t callbacks */
-enum {
-    FDB_MSG_TRAN_BEGIN = 0,
-    FDB_MSG_TRAN_PREPARE = 1,
-    FDB_MSG_TRAN_COMMIT = 2,
-    FDB_MSG_TRAN_ROLLBACK = 3,
-    FDB_MSG_TRAN_RC = 4,
-
-    FDB_MSG_CURSOR_OPEN = 5,
-    FDB_MSG_CURSOR_CLOSE = 6,
-    FDB_MSG_CURSOR_FIND = 7,
-    FDB_MSG_CURSOR_FIND_LAST = 8,
-    FDB_MSG_CURSOR_FIRST = 9,
-    FDB_MSG_CURSOR_LAST = 10,
-    FDB_MSG_CURSOR_NEXT = 11,
-    FDB_MSG_CURSOR_PREV = 12,
-
-    FDB_MSG_DATA_ROW = 13,
-    FDB_MSG_DATA_RC = 14,
-
-    FDB_MSG_RUN_SQL = 15,
-
-    FDB_MSG_INSERT = 16,
-    FDB_MSG_DELETE = 17,
-    FDB_MSG_UPDATE = 18,
-
-    FDB_MSG_HBEAT = 19,
-
-    /* OPs for partial indexes */
-    FDB_MSG_INSERT_PI = 20,
-    FDB_MSG_DELETE_PI = 21,
-    FDB_MSG_UPDATE_PI = 22,
-
-    FDB_MSG_INDEX = 23,
-
-    FDB_MSG_MAX_OP
-};
-
-enum { FD_MSG_TYPE = 0x0fff, FD_MSG_FLAGS_ISUUID = 0x1000 };
-
-struct fdb_msg_header {
-    int type;
-};
-typedef struct fdb_msg_header fdb_msg_header_t;
-
-struct prefilt_sql {
-    int l_incl; /* length of bitmap */
-    char *incl; /* bitmap of ondisk included columns */
-    int l_expr; /* length of expression */
-    char *expr; /* serialized expression */
-};
-
-struct fdb_msg_cursor_open {
-    fdb_msg_header_t type; /* FDB_MSG_CURSOR_OPEN */
-    char *cid;             /* cursor id */
-    char *tid;             /* transaction id, create one if not existing */
-    int flags;             /* feature-based */
-    int rootpage;          /* which btree */
-    int version;           /* which version of schema */
-    int seq;               /* used for ordering of certain operations in
-                              transactional cursors (see fdb_msg_tran_t) */
-    uuid_t ciduuid;
-    uuid_t tiduuid;
-    int srcpid;     /* pid of the source */
-    int srcnamelen; /* hostname of the source */
-    char *srcname;
-};
-typedef struct fdb_msg_cursor_open fdb_msg_cursor_open_t;
-
-struct fdb_msg_cursor_close {
-    fdb_msg_header_t type; /* FDB_MSG_CURSOR_CLOSE */
-    char *cid;             /* cursor id */
-    char *tid;             /* transaction id */
-    uuid_t ciduuid;
-    uuid_t tiduuid;
-    int seq; /* used for ordering of certain operations in
-                transactional cursors (see fdb_msg_tran_t); */
-};
-typedef struct fdb_msg_cursor_close fdb_msg_cursor_close_t;
-
-struct fdb_msg_cursor_move {
-    fdb_msg_header_t type; /* FDB_MSG_CURSOR_FIND, .... */
-    char *cid;             /* cursor id */
-    uuid_t ciduuid;
-};
-
-typedef struct fdb_msg_cursor_move fdb_msg_cursor_move_t;
-
-struct fdb_msg_tran {
-    fdb_msg_header_t type;      /* FDB_MSG_TRAN_BEGIN, ... */
-    char *tid;                  /* transaction id */
-    enum transaction_level lvl; /* TRANLEVEL_SOSQL & co. */
-    int flags;                  /* extensions */
-    uuid_t tiduuid;
-    int seq; /* sequencing tran begin/commit/rollback, writes, cursor open/close
-                */
-};
-typedef struct fdb_msg_tran fdb_msg_tran_t;
-
-struct fdb_msg_tran_rc {
-    fdb_msg_header_t type; /* FDB_MS_TRAN_RC */
-    char *tid;             /* transaction id */
-    int rc;                /* result code */
-    int errstrlen;         /* error string length */
-    uuid_t tiduuid;
-    char *errstr; /* error string, if any */
-};
-typedef struct fdb_msg_tran_rc fdb_msg_tran_rc_t;
-
-struct fdb_msg_data_row {
-    fdb_msg_header_t type;    /* FDB_MSG_DATA_ROW */
-    char *cid;                /* id of the cursor */
-    int rc;                   /* results of the search */
-    unsigned long long genid; /* genid */
-    int datalen; /* NET FORMAT = data length: high 16 bits; datacopy length: low
-                    16 bits*/
-    int datacopylen; /* datacopylen after extraction */
-    char *data;      /* row(including datacopy) or error string */
-    char *datacopy;  /* after extraction */
-    uuid_t ciduuid;
-};
-typedef struct fdb_msg_data_row fdb_msg_data_row_t;
-
-struct fdb_msg_cursor_find {
-    fdb_msg_header_t type; /* FDB_MSG_CURSOR_FIND */
-    char *cid;             /* id of the cursor */
-    int keylen;            /* keylen, serialized sqlite */
-    char *key;             /* key, serialized sqlite */
-    uuid_t ciduuid;
-};
-typedef struct fdb_msg_cursor_find fdb_msg_cursor_find_t;
-
-#define FDB_RUN_SQL
-
-struct fdb_msg_run_sql {
-    fdb_msg_header_t type;    /* FDB_MSG_RUN_SQL */
-    char *cid;                /* cursor id */
-    int version;              /* schema version */
-    enum run_sql_flags flags; /* flags changing the remote behaviour */
-    int sqllen;               /* len of sql query */
-    char *sql;                /* sql query */
-    int keylen; /* keylen used for end trimming  NOT USED RIGHT NOW!*/
-    char *key;  /* key ised for end trimming  NOT USED RIGHT NOW! */
-    uuid_t ciduuid;
-};
-typedef struct fdb_msg_run_sql fdb_msg_run_sql_t;
-
-struct fdb_msg_insert {
-    fdb_msg_header_t type;       /* FDB_MSG_INSERT */
-    char *cid;                   /* cursor id */
-    int version;                 /* schema version */
-    int rootpage;                /* which btree I am inserting into */
-    unsigned long long genid;    /* genid */
-    unsigned long long ins_keys; /* indexes to insert */
-    int datalen;                 /* length of sqlite row, see below */
-    int seq;                     /* transaction sequencing */
-    uuid_t ciduuid;
-    char *data; /* sqlite generated row from MakeRecord, serialized */
-    char *tblname; /* tblname matching rootpage */
-};
-typedef struct fdb_msg_insert fdb_msg_insert_t;
-
-struct fdb_msg_index {
-    fdb_msg_header_t type;    /* FDB_MSG_INDEX */
-    char *cid;                /* cursor id */
-    int version;              /* schema version */
-    int rootpage;             /* which btree I am inserting into */
-    unsigned long long genid; /* genid */
-    int is_delete;            /* 1 for delete, 0 for add */
-    int ixnum;                /* index number */
-    int ixlen;                /* length of sqlite index row, see below */
-    int seq;                  /* transaction sequencing */
-    uuid_t ciduuid;
-    char *ix; /* sqlite generated index, serialized */
-};
-typedef struct fdb_msg_index fdb_msg_index_t;
-
-struct fdb_msg_delete {
-    fdb_msg_header_t type;       /* FDB_MSG_DELETE */
-    char *cid;                   /* cursor id */
-    int version;                 /* schema version */
-    int rootpage;                /* which btree I am deleting from */
-    unsigned long long genid;    /* genid */
-    unsigned long long del_keys; /* indexes to delete */
-    uuid_t ciduuid;
-    int seq; /* transaction sequencing */
-    char *tblname; /* tblname matching rootpage */
-};
-typedef struct fdb_msg_delete fdb_msg_delete_t;
-
-struct fdb_msg_update {
-    fdb_msg_header_t type;       /* FDB_MSG_UPDATE */
-    char *cid;                   /* cursor id */
-    int version;                 /* schema version */
-    int rootpage;                /* which btree I am inserting into */
-    unsigned long long oldgenid; /* oldgenid */
-    unsigned long long genid;    /* genid */
-    unsigned long long ins_keys; /* indexes to insert */
-    unsigned long long del_keys; /* indexes to delete */
-    int datalen;                 /* length of sqlite row, see below */
-    int seq;                     /* transaction sequencing */
-    uuid_t ciduuid;
-    char *data; /* sqlite generated row from MakeRecord, serialized */
-    char *tblname; /* tblname matching rootpage */
-};
-typedef struct fdb_msg_update fdb_msg_update_t;
-
-struct fdb_msg_hbeat {
-    fdb_msg_header_t type;    /* FDB_MSG_UPDATE */
-    char *tid;                /* tran id */
-    struct timespec timespec; /* when was this hbeat sent */
-    uuid_t tiduuid;
-};
-typedef struct fdb_msg_hbeat fdb_msg_hbeat_t;
-
-union fdb_msg {
-    fdb_msg_header_t hd;
-    fdb_msg_cursor_open_t co;
-    fdb_msg_cursor_close_t cc;
-    fdb_msg_cursor_move_t cm;
-    fdb_msg_tran_t tr;
-    fdb_msg_tran_rc_t rc;
-    fdb_msg_data_row_t dr;
-    fdb_msg_cursor_find_t cf;
-    fdb_msg_run_sql_t sq;
-    fdb_msg_insert_t in;
-    fdb_msg_delete_t de;
-    fdb_msg_update_t up;
-    fdb_msg_index_t ix;
-    fdb_msg_hbeat_t hb;
-};
-
-static int fdb_msg_read_message(SBUF2 *sb, fdb_msg_t *msg,
-                                enum recv_flags flags);
 static int fdb_msg_write_message(SBUF2 *sb, fdb_msg_t *msg, int flush);
-static void fdb_msg_print_message(SBUF2 *sb, fdb_msg_t *msg, char *prefix);
-static void fdb_msg_print_message_uuid(SBUF2 *sb, fdb_msg_t *msg, char *prefix);
-
-typedef struct {
-    struct sqlclntstate *clnt;
-    struct sql_thread *thd;
-    int isuuid;
-    int flags;
-} svc_callback_arg_t;
-
-typedef int (*fdb_svc_callback_t)(SBUF2 *sb, fdb_msg_t *msg,
-                                  svc_callback_arg_t *arg);
 
 int fdb_remcur_trans_begin(SBUF2 *sb, fdb_msg_t *msg, svc_callback_arg_t *arg);
 int fdb_remcur_trans_prepare(SBUF2 *sb, fdb_msg_t *msg,
@@ -573,361 +326,8 @@ done:
 
 int fdb_msg_size(void) { return sizeof(fdb_msg_t); }
 
-static int _check_code_release(SBUF2 *sb, char *cid, int code_release,
-                               int isuuid)
-{
-    char errstr[256];
-    int errval;
-    int rc;
-
-    code_release = fdb_ver_decoded(code_release);
-
-    /* lets make sure we ask for sender to downgrade if its code is too new */
-    if (unlikely(code_release > FDB_VER)) {
-
-        snprintf(errstr, sizeof(errstr), "%d protocol %d too high", FDB_VER,
-                 code_release);
-        errval = FDB_ERR_FDB_VERSION;
-
-        /* we need to send back a rc code */
-        rc = fdb_svc_sql_row(sb, cid, errstr, strlen(errstr) + 1, errval,
-                             isuuid);
-        if (rc) {
-            logmsg(LOGMSG_ERROR, "%s: fdb_send_rc failed rc=%d\n", __func__, rc);
-        }
-
-        return -1;
-    }
-
-    return 0;
-}
-
-static int handle_remsql_session(SBUF2 *sb, struct dbenv *dbenv)
-{
-    fdb_msg_cursor_open_t open_msg;
-    fdb_msg_t msg;
-    int rc = 0;
-    svc_callback_arg_t arg;
-    int flags;
-
-    bzero(&msg, sizeof(msg));
-
-    rc = fdb_msg_read_message(sb, &msg, 0);
-    if (rc) {
-        logmsg(LOGMSG_ERROR, "%s: failed to handle remote cursor request rc=%d\n",
-                __func__, rc);
-        return rc;
-    }
-
-    flags = msg.hd.type & ~FD_MSG_TYPE;
-    msg.hd.type &= FD_MSG_TYPE;
-
-    if (msg.hd.type != FDB_MSG_CURSOR_OPEN) {
-        logmsg(LOGMSG_ERROR, 
-                "%s: received wrong packet type=%d, expecting cursor open\n",
-                __func__, msg.hd.type);
-        return -1;
-    }
-
-    memcpy(&open_msg, &msg, sizeof open_msg);
-
-    /* check and protect against newer versions */
-    if (_check_code_release(sb, open_msg.cid, open_msg.rootpage,
-                            flags & FD_MSG_FLAGS_ISUUID)) {
-        logmsg(LOGMSG_ERROR, "PROTOCOL TOO NEW %d, asking to downgrade to %d\n",
-                fdb_ver_decoded(open_msg.rootpage), FDB_VER);
-        return 0;
-    }
-
-    while (1) {
-
-        arg.isuuid = flags & FD_MSG_FLAGS_ISUUID;
-
-        if (gbl_fdb_track) {
-            if (arg.isuuid) {
-                fdb_msg_print_message_uuid(sb, &msg, "received msg");
-            } else
-                fdb_msg_print_message(sb, &msg, "received msg");
-        }
-
-        rc = callbacks[msg.hd.type](sb, &msg, &arg);
-
-        if (msg.hd.type == FDB_MSG_CURSOR_CLOSE) {
-            break;
-        }
-        if (rc != 0) {
-            int rc2;
-
-            rc2 = fdb_svc_cursor_close(open_msg.cid,
-                                       flags & FD_MSG_FLAGS_ISUUID, &arg.clnt);
-            if (rc2) {
-                logmsg(LOGMSG_ERROR, "%s: fdb_svc_cursor_close failed rc=%d\n",
-                        __func__, rc2);
-            }
-            break;
-        }
-
-        /*fprintf(stderr, "XYXY %llu calling recv message\n",
-         * osql_log_time());*/
-        rc = fdb_msg_read_message(sb, &msg, 0);
-        if (rc) {
-            logmsg(LOGMSG_ERROR, 
-                    "%s: failed to handle remote cursor request rc=%d\n",
-                    __func__, rc);
-            break;
-        }
-
-        flags = msg.hd.type & ~FD_MSG_TYPE;
-        msg.hd.type &= FD_MSG_TYPE;
-    }
-
-    fdb_msg_clean_message(&msg);
-
-    return rc;
-}
-
-int handle_remsql(comdb2_appsock_arg_t *arg)
-{
-    struct dbenv *dbenv;
-    struct sbuf2 *sb;
-    char line[128];
-    int rc = FDB_NOERR;
-    static uint64_t old = 0ULL;
-    uint64_t now;
-    uint64_t then;
-
-    dbenv = arg->dbenv;
-    sb = arg->sb;
-
-    /* We will rely on socket drop to determine the end of connection. */
-    sbuf2settimeout(sb, 0, 0);
-
-    if (gbl_fdb_track_times) {
-        now = gettimeofday_ms();
-        logmsg(LOGMSG_USER, "RRRRRR start now=%lu\n", now);
-    }
-
-    while (1) {
-        if (gbl_fdb_track_times) {
-            now = gettimeofday_ms();
-        }
-
-        rc = handle_remsql_session(sb, dbenv);
-        if (gbl_fdb_track)
-            logmsg(LOGMSG_USER, "%lu: %s: executed session rc=%d\n",
-                   pthread_self(), __func__, rc);
-
-        if (gbl_fdb_track_times) {
-            then = gettimeofday_ms();
-
-            if (old == 0ULL) {
-                logmsg(LOGMSG_USER, "RRRRRR now=%lu 0 %lu\n", now, then - now);
-            } else {
-                logmsg(LOGMSG_USER, "RRRRRR now=%lu delta=%lu %lu\n", now,
-                       now - old, then - now);
-            }
-            old = now;
-        }
-
-        if (rc == FDB_NOERR) {
-            /* we need to read the header again, waiting here */
-            rc = sbuf2gets(line, sizeof(line), sb);
-            if (rc != strlen("remsql\n")) {
-                if (rc != -1)
-                    logmsg(LOGMSG_ERROR,
-                           "%s: received wrong request! rc=%d: %s\n", __func__,
-                           rc, line);
-                rc = FDB_NOERR;
-                break;
-            }
-            /* execute next session */
-            continue;
-        } else {
-            break;
-        }
-    }
-    if (gbl_fdb_track)
-        logmsg(LOGMSG_USER, "%lu: %s: done processing\n", pthread_self(),
-               __func__);
-
-    return rc;
-}
-
-int handle_remtran(comdb2_appsock_arg_t *arg)
-{
-    struct sbuf2 *sb;
-    fdb_msg_tran_t open_msg;
-    fdb_msg_t msg;
-    int rc = 0;
-    svc_callback_arg_t svc_cb_arg = {0};
-
-    sb = arg->sb;
-
-    bzero(&msg, sizeof(msg));
-
-    /* This does insert on behalf of an sql transaction */
-    svc_cb_arg.thd = start_sql_thread();
-
-    rc = fdb_msg_read_message(sb, &msg, 0);
-    if (rc) {
-        logmsg(LOGMSG_ERROR,
-               "%s: failed to handle remote cursor request rc=%d\n", __func__,
-               rc);
-        return rc;
-    }
-
-    if ((msg.hd.type & FD_MSG_TYPE) != FDB_MSG_TRAN_BEGIN) {
-        logmsg(LOGMSG_ERROR,
-               "%s: received wrong packet type=%d, expecting tran begin\n",
-               __func__, msg.hd.type);
-        return -1;
-    }
-
-    memcpy(&open_msg, &msg, sizeof open_msg);
-    open_msg.tid = (char *)open_msg.tiduuid;
-    uuidstr_t us;
-    comdb2uuidstr((unsigned char *)open_msg.tid, us);
-
-    /* TODO: review the no-timeout transaction later on */
-    if (gbl_notimeouts) {
-        sbuf2settimeout(sb, 0, 0);
-        /*   net_add_watch(sb, 0, 0); */
-    }
-
-    while (1) {
-        if (gbl_fdb_track) {
-            fdb_msg_print_message(sb, &msg, "received msg");
-        }
-
-        svc_cb_arg.isuuid = (msg.hd.type & FD_MSG_FLAGS_ISUUID);
-
-        rc = callbacks[msg.hd.type & FD_MSG_TYPE](sb, &msg, &svc_cb_arg);
-
-        if ((msg.hd.type & FD_MSG_TYPE) == FDB_MSG_TRAN_COMMIT ||
-            (msg.hd.type & FD_MSG_TYPE) == FDB_MSG_TRAN_ROLLBACK ||
-            (msg.hd.type & FD_MSG_TYPE) ==
-                FDB_MSG_TRAN_RC /* this should be actuall the only case,
-                  since we reuse the buffer to send back results */
-            ) {
-            if ((msg.hd.type & FD_MSG_TYPE) == FDB_MSG_TRAN_COMMIT ||
-                (msg.hd.type & FD_MSG_TYPE) == FDB_MSG_TRAN_ROLLBACK)
-                abort();
-
-            break;
-        }
-        if (rc != 0) {
-            int rc2;
-        clear:
-            rc2 = fdb_svc_trans_rollback(
-                open_msg.tid, open_msg.lvl, svc_cb_arg.clnt,
-                svc_cb_arg.clnt->dbtran.dtran->fdb_trans.top->seq);
-            if (rc2) {
-                logmsg(LOGMSG_ERROR,
-                       "%s: fdb_svc_trans_rollback failed rc=%d\n", __func__,
-                       rc2);
-            }
-            break;
-        }
-
-        /*fprintf(stderr, "XYXY %llu calling recv message\n",
-         * osql_log_time());*/
-        rc = fdb_msg_read_message(sb, &msg, svc_cb_arg.flags);
-        if (rc) {
-            logmsg(LOGMSG_ERROR,
-                   "%s: failed to handle remote cursor request rc=%d\n",
-                   __func__, rc);
-            goto clear;
-        }
-    }
-
-    if (gbl_expressions_indexes) {
-        free(svc_cb_arg.clnt->idxInsert);
-        free(svc_cb_arg.clnt->idxDelete);
-        svc_cb_arg.clnt->idxInsert = svc_cb_arg.clnt->idxDelete = NULL;
-    }
-
-    reset_clnt(svc_cb_arg.clnt, NULL, 0);
-
-    done_sql_thread();
-
-    pthread_mutex_destroy(&svc_cb_arg.clnt->wait_mutex);
-    pthread_cond_destroy(&svc_cb_arg.clnt->wait_cond);
-    pthread_mutex_destroy(&svc_cb_arg.clnt->write_lock);
-    pthread_mutex_destroy(&svc_cb_arg.clnt->dtran_mtx);
-
-    free(svc_cb_arg.clnt);
-    svc_cb_arg.clnt = NULL;
-
-    return rc;
-}
-
-int handle_remcur(comdb2_appsock_arg_t *arg)
-{
-    struct sbuf2 *sb;
-    fdb_msg_cursor_open_t open_msg;
-    fdb_msg_t msg;
-    int rc = 0;
-    svc_callback_arg_t svc_cb_arg = {0};
-    int flags;
-
-    sb = arg->sb;
-
-    bzero(&msg, sizeof(msg));
-
-    logmsg(LOGMSG_DEBUG, "Received remcu request\n");
-
-    rc = fdb_msg_read_message(sb, &msg, 0);
-    if (rc) {
-        logmsg(LOGMSG_ERROR,
-               "%s: failed to handle remote cursor request rc=%d\n", __func__,
-               rc);
-        return rc;
-    }
-
-    flags = msg.hd.type & ~FD_MSG_TYPE;
-    msg.hd.type &= FD_MSG_TYPE;
-
-    if (msg.hd.type != FDB_MSG_CURSOR_OPEN) {
-        logmsg(LOGMSG_ERROR,
-               "%s: received wrong packet type=%d, expecting cursor open\n",
-               __func__, msg.hd.type);
-        return -1;
-    }
-
-    memcpy(&open_msg, &msg, sizeof open_msg);
-
-    while (1) {
-        if (gbl_fdb_track) {
-            fdb_msg_print_message(sb, &msg, "received msg");
-        }
-
-        rc = callbacks[msg.hd.type](sb, &msg, &svc_cb_arg);
-        if (msg.hd.type == FDB_MSG_CURSOR_CLOSE) {
-            break;
-        }
-        if (rc != 0) {
-            int rc2;
-            rc2 = fdb_svc_cursor_close(open_msg.cid,
-                                       flags & FD_MSG_FLAGS_ISUUID, NULL);
-            if (rc2 && !rc)
-                rc = rc2;
-            break;
-        }
-
-        rc = fdb_msg_read_message(sb, &msg, 0);
-        if (rc) {
-            logmsg(LOGMSG_ERROR,
-                   "%s: failed to handle remote cursor request rc=%d\n",
-                   __func__, rc);
-            break;
-        }
-
-        flags = msg.hd.type & ~FD_MSG_TYPE;
-        msg.hd.type &= FD_MSG_TYPE;
-    }
-
-    return rc;
-}
+int fdb_msg_read_message(SBUF2 *sb, fdb_msg_t *msg,
+                         enum recv_flags flags);
 
 int fdb_recv_row(fdb_msg_t *msg, char *cid, SBUF2 *sb)
 {
@@ -1220,8 +620,8 @@ static void fdb_msg_prepare_message(fdb_msg_t *msg)
 }
 
 /* stuff comes in network endian fomat */
-static int fdb_msg_read_message(SBUF2 *sb, fdb_msg_t *msg,
-                                enum recv_flags flags)
+int fdb_msg_read_message(SBUF2 *sb, fdb_msg_t *msg,
+                         enum recv_flags flags)
 {
     int rc;
     unsigned long long lltmp;
@@ -1927,7 +1327,7 @@ static const char *__tran_2_str(enum transaction_level lvl)
     return "???";
 }
 
-static void fdb_msg_print_message_uuid(SBUF2 *sb, fdb_msg_t *msg, char *prefix)
+void fdb_msg_print_message_uuid(SBUF2 *sb, fdb_msg_t *msg, char *prefix)
 {
     int rc;
     int i;
@@ -2100,7 +1500,7 @@ static void fdb_msg_print_message_uuid(SBUF2 *sb, fdb_msg_t *msg, char *prefix)
     }
 }
 
-static void fdb_msg_print_message(SBUF2 *sb, fdb_msg_t *msg, char *prefix)
+void fdb_msg_print_message(SBUF2 *sb, fdb_msg_t *msg, char *prefix)
 {
     int rc;
     int i;
