+++ db/fdb_fend.c
@@ -3956,135 +3956,6 @@ int fdb_is_sqlite_stat(fdb_t *fdb, int rootpage)
            0;
 }
 
-static int _is_tablename_unique(const char *name)
-{
-    int i;
-    int llen = strlen(name);
-
-    for (i = 0; i < thedb->num_dbs; i++) {
-        if (llen != strlen(thedb->dbs[i]->tablename))
-            continue;
-        if (strncasecmp(thedb->dbs[i]->tablename, name, llen) == 0)
-            return -1;
-    }
-
-    return 0;
-}
-
-int fdb_alias_command(comdb2_appsock_arg_t *arg)
-{
-    struct dbenv *dbenv;
-    struct sbuf2 *sb;
-    char *op = NULL;
-    char *aliasname = NULL;
-    char *url = NULL;
-    char *errstr = NULL;
-    char *tok;
-    char *line;
-    int llen;
-    int ltok = 0;
-    int st = 0;
-    int rc = -1;
-
-    dbenv = arg->dbenv;
-    sb = arg->sb;
-    line = arg->cmdline;
-    llen = strlen(line);
-
-    if (dbenv->master != gbl_mynode) {
-        sbuf2printf(sb, "!master swinged, now on %s, please rerun\n",
-                    thedb->master);
-        sbuf2printf(sb, "FAILED\n");
-        return APPSOCK_RETURN_CONT;
-    }
-
-    tok = segtok(line, llen, &st, &ltok);
-    assert((strncmp(tok, "alias", ltok) == 0));
-
-    tok = segtok(line, llen, &st, &ltok);
-    if (ltok == 0) {
-    usage:
-        sbuf2printf(sb, "!Usage: [alias set ALIASNAME URL|alias get ALIASNAME| "
-                        "alias rem ALIASNAME]\n");
-        sbuf2printf(sb, "FAILED\n");
-
-        if (op)
-            free(op);
-        if (aliasname)
-            free(aliasname);
-        if (url)
-            free(url);
-
-        arg->error = FDB_ERR_GENERIC;
-        return APPSOCK_RETURN_ERR;
-    }
-
-    op = tokdup(tok, ltok);
-    if (!op) {
-
-        arg->error = FDB_ERR_MALLOC;
-        return APPSOCK_RETURN_ERR;
-    }
-
-    tok = segtok(line, llen, &st, &ltok);
-    if (ltok == 0)
-        goto usage;
-    aliasname = tokdup(tok, ltok);
-
-    rc = _is_tablename_unique(aliasname);
-    if (rc) {
-        char msg[256];
-        snprintf(msg, sizeof(msg),
-                 "alias \"%s\" exists as table name, must be unique",
-                 aliasname);
-        sbuf2printf(sb, "!%s\n", msg);
-        sbuf2printf(sb, "FAILED\n");
-        arg->error = FDB_ERR_GENERIC;
-        return APPSOCK_RETURN_ERR;
-    }
-
-    if (strncasecmp(op, "set", 3) == 0) {
-        tok = segtok(line, llen, &st, &ltok);
-        if (ltok == 0)
-            goto usage;
-        url = tokdup(tok, ltok);
-
-        rc = llmeta_set_tablename_alias(NULL, aliasname, url, &errstr);
-    } else if (strncasecmp(op, "get", 3) == 0) {
-        url = llmeta_get_tablename_alias(aliasname, &errstr);
-        rc = (url == 0);
-
-        if (rc == 0) {
-            sbuf2printf(sb, ">%s\n", url);
-        }
-    } else if (strncasecmp(op, "rem", 3) == 0) {
-        rc = llmeta_rem_tablename_alias(aliasname, &errstr);
-    } else {
-        goto usage;
-    }
-
-    if (op)
-        free(op);
-    if (aliasname)
-        free(aliasname);
-    if (url)
-        free(url);
-
-    if (rc) {
-        sbuf2printf(sb, "!%s\n", (errstr) ? errstr : "no string");
-        sbuf2printf(sb, "FAILED\n");
-
-        if (errstr) {
-            free(errstr);
-        }
-    } else {
-        sbuf2printf(sb, "SUCCESS\n");
-    }
-
-    arg->error = rc;
-    return (rc) ? APPSOCK_RETURN_ERR : APPSOCK_RETURN_OK;
-}
-
 char *fdb_get_alias(const char **p_tablename)
 {
     char *errstr = NULL;
