+++ plugin/remsql/plugin.h.in
@@ -1,0 +1,47 @@
+struct comdb2_plugin @PLUGIN_SYM@[] = {
+    {
+        "remcur",              /* Plugin identifier */
+        "remcur plugin",       /* Plugin description */
+        COMDB2_PLUGIN_APPSOCK, /* Plugin type */
+        1,                     /* Plugin version */
+        1,                     /* Plugin interface version */
+        0,                     /* Plugin flags */
+        remsql_init,           /* Initialization function */
+        remsql_destroy,        /* Destroy function */
+        &remcur_plugin         /* Plugin-specific data */
+    },
+    {
+        "remsql",              /* Plugin identifier */
+        "remsql plugin",       /* Plugin description */
+        COMDB2_PLUGIN_APPSOCK, /* Plugin type */
+        1,                     /* Plugin version */
+        1,                     /* Plugin interface version */
+        0,                     /* Plugin flags */
+        remsql_init,           /* Initialization function */
+        remsql_destroy,        /* Destroy function */
+        &remsql_plugin         /* Plugin-specific data */
+    },
+    {
+
+        "remtran",             /* Plugin identifier */
+        "remtran plugin",      /* Plugin description */
+        COMDB2_PLUGIN_APPSOCK, /* Plugin type */
+        1,                     /* Plugin version */
+        1,                     /* Plugin interface version */
+        0,                     /* Plugin flags */
+        remsql_init,           /* Initialization function */
+        remsql_destroy,        /* Destroy function */
+        &remtran_plugin        /* Plugin-specific data */
+    },
+    {
+        "alias",               /* Plugin identifier */
+        "alias plugin",        /* Plugin description */
+        COMDB2_PLUGIN_APPSOCK, /* Plugin type */
+        1,                     /* Plugin version */
+        1,                     /* Plugin interface version */
+        0,                     /* Plugin flags */
+        remsql_init,           /* Initialization function */
+        remsql_destroy,        /* Destroy function */
+        &alias_plugin          /* Plugin-specific data */
+    },
+    {0, 0, 0, 0, 0, 0, 0, 0, 0}};
