+++ schemachange/schemachange.c
@@ -1153,137 +1153,6 @@ error_prelock:
     return xerr->errval;
 }
 
-/* wrapper handling transaction */
-static int do_partition(timepart_views_t *views, const char *name,
-                        const char *cmd, struct errstat *err)
-{
-    struct ireq iq;
-    void *tran = NULL;
-    int rc;
-    int irc = 0;
-    int bdberr;
-    /*
-        init_fake_ireq(thedb, &iq);
-        iq.use_handle = thedb->bdb_env;
-
-        rc = trans_start(&iq, NULL, &tran);
-        if (rc)
-        {
-            err->errval = VIEW_ERR_GENERIC;
-            snprintf(err->errstr, sizeof(err->errstr), ">%s -- trans_start
-       rc:%d\n", __func__, rc);
-            return rc;
-        }
-    */
-    rc = views_do_partition(tran, views, name, cmd, err);
-    /*
-        if(rc)
-        {
-            irc = trans_abort(&iq, tran);
-        }
-        else
-        {
-            irc = trans_commit(&iq, tran, gbl_mynode);
-        }
-        if(irc)
-        {
-            err->errval = VIEW_ERR_GENERIC;
-            snprintf(err->errstr, sizeof(err->errstr), ">%s -- trans_%s
-       rc:%d\n",
-                    __func__, (rc)?"abort":"commit", irc);
-        }
-
-        tran = NULL;
-    */
-    if (!rc && !irc) {
-        bdberr = 0;
-        irc = bdb_llog_views(thedb->bdb_env, (char *)name, 1, &bdberr);
-        if (irc) {
-            logmsg(LOGMSG_ERROR, "%s -- bdb_llog_views rc:%d bdberr:%d\n",
-                   __func__, irc, bdberr);
-        }
-    }
-
-    return rc;
-}
-
-int handle_partition(comdb2_appsock_arg_t *arg)
-{
-#define CHUNK 512
-    struct errstat xerr = {0};
-    struct sbuf2 *sb;
-    char line[CHUNK];
-    char *viewname;
-    char *cmd;
-    int len;
-    int used;
-    int alloc;
-    int rc;
-
-    sb = arg->sb;
-
-    if ((rc = sbuf2gets(line, sizeof(line), sb)) < 0) {
-        logmsg(LOGMSG_ERROR, "%s -- sbuf2gets rc: %d\n", __func__, rc);
-        arg->error = 1;
-        return APPSOCK_RETURN_ERR;
-    }
-    viewname = strdup(line);
-    if (viewname[strlen(viewname) - 1] == '\n') {
-        viewname[strlen(viewname) - 1] = '\0';
-    }
-
-    /* Now, read new schema */
-    used = alloc = 0;
-    cmd = NULL;
-    while ((rc = sbuf2gets(line, sizeof(line), sb)) > 0) {
-        if (strcmp(line, ".\n") == 0) {
-            /* we know we have added a newline to mark end of schema,
-               delete that */
-            if (cmd && used > 0 && cmd[used - 1] == '\n') {
-                cmd[used - 1] = '\0';
-                len--;
-            }
-            break;
-        }
-
-        len = strlen(line) + 1; /* include \0 terminator */
-        if ((alloc - used) < len) {
-            if (len < CHUNK) {
-                alloc = alloc + CHUNK;
-            } else {
-                alloc = alloc + len;
-            }
-            cmd = realloc(cmd, alloc);
-            if (!cmd) {
-                logmsg(LOGMSG_ERROR,
-                       "appsock_schema_change_inner: out of memory buf=%d\n",
-                       alloc);
-                sbuf2printf(sb, "!out of memory reading schema\n");
-                sbuf2printf(sb, "FAILED\n");
-                arg->error = 1;
-                return APPSOCK_RETURN_ERR;
-            }
-        }
-        memcpy(cmd + used, line, len);
-        used += len - 1; /* don't add \0 terminator */
-    }
-
-    /* do the work */
-    wrlock_schema_lk();
-    rc = do_partition(thedb->timepart_views, viewname, cmd, &xerr);
-    unlock_schema_lk();
-
-    if (rc == 0)
-        sbuf2printf(sb, "SUCCESS\n");
-    else
-    out:
-    sbuf2printf(sb, "FAILED rc %d err \"%s\"\n", xerr.errval, xerr.errstr);
-
-    sbuf2flush(sb);
-
-    return APPSOCK_RETURN_OK;
-}
-
 /* shortcut for running table upgrade in a schemachange shell */
 int start_table_upgrade(struct dbenv *dbenv, const char *tbl,
                         unsigned long long genid, int full, int partial,
